Investigation environment initialised.
I have created the environment, as well as the repository for the this investigation. I have set up the README for future use, as well as imported the dataset (names.txt), and  created both the main.py Python executable and a PowerPoint presentation called slides.pptx. I have also decided to give this project the unique name of MonkeyMLM, as I wish to work on this project and evolve it even after the due date of the assignment.