Section 2.1 Completed
I have completed section 2.1 by integrating a function to read the file, and a function that returns the longest name(s) and the shortest name(s). I achieved this by using a for statement to skip through the document and locate all the words with the longest length of characters, and the same with the smallest, and append them to two different lists.