Section 3.1 Completed
Hello commit we meet again. I have completed work up to and including 3.1 today. Honestly, no sweat 😎. To make the rigged coin flip, I made the function roll one of the numbers between 1 to 10 (inclusive) and based on the number that was rolled it would give the calculate whether heads or tails was rolled. To test its correctness, I ran a unit test that ran the function many times and calculated the percentage of each value being picked. I followed the same principal with the spinner function and the same way to test its accuracy.

Godspeed Future Dylan 🫡
(not too long left)