Section 3.2 Completed
Hallo. Today I completed sections up to and including 3.2. I might be him 😊. All I've done these last two days was create the necessary functions for the name generation. The first function calculates the probability of a pair being generated based on its frequency based on the total frequency of all pairs that begin with that letter. The second function unpacks the data and uses the probabilities we just made to pick a pair. The final function feeds input into the second function to generate a name of a random length and returns that string. The returned string is the assumed generated name.

P.S. sorry for the self-glazing I'll stop