Updates to the repository
I noticed that there are a few file in the repository that I don't want, purely because they aren't necessary to the functionality of the investigation, so I removed the previously committed '.idea' folder and its contents. I am aware of the .DS_Store file, I don't know what it is and I am not able to remove it using my Macbook as I am not sure what the setting is to view hidden files/folders.